phase1067-lint-google-rule-site-fixes(engine+samples)
Sprint-1 lint queue, wave 1 of the 12 deferred-rule probe: full-tree
site fixes for three google rules ahead of their promotion.

google-runtime-int (12 sites): test counters long long -> std::int64_t;
hello_engine sleep cast -> std::int64_t; strtol/lround/Khronos-ABI
sites annotated NOLINT with the C-ABI rationale (the return types are
fixed by the C standard / khrplatform.h, not ours to change).

google-build-using-namespace (15 sites, all tests): function-scope and
file-scope using-directives replaced with explicit using-declarations
(UDL operator""_deg/_rad imports in test_math; symbol lists in the four
file-scope test TUs; file-scope declaration blocks for the repeated
function-scope directives in test_material + test_render -- the
test_material block sits after the mid-file BrdfLut.hpp include that
declares the symbols).

google-explicit-constructor (16 sites): FixedString(string_view),
CompactLocation, AckWindowChannel, test Node ctor marked explicit (all
call sites already brace-init); ArgvBuilder's initializer_list ctor
DROPS explicit (the check's reverse arm); Json::Value converting ctors
and Ref<T> get NOLINT blocks -- implicit conversion is the documented
design there (nlohmann-style JSON value type; std::reference_wrapper
semantics).

Build clean, 262/262 ctest PASS. Promotion of the three rules lands
with the .clang-tidy update after the remaining two waves
(llvm-namespace-comment autofix + google-default-arguments NVI).

+++ samples/engine/hello_engine/main.cpp
@@ -6745,7 +6745,7 @@ struct HelloEngineApp::EngineState
                 const auto v = id.value();
                 std::this_thread::sleep_for(
                     std::chrono::milliseconds(
-                        120 + static_cast<long long>(v % 5) * 80));
+                        120 + static_cast<std::int64_t>(v % 5) * 80));
                 const bool ok = (v % 17 != 0);
                 if (ok) streamer_completed.fetch_add(1, std::memory_order_relaxed);
                 else    streamer_failed.fetch_add(1, std::memory_order_relaxed);
